+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 * Added modulo.
 * Fixed an issue where the compiler would hang on some loops with optimizations turned on.
 * Added `continue` and `break`.
+* Added `unsafe_string`.
+* Added `tag` and `team` target selector arguments.